+++ junction/proposals/urls.py
@@ -23,6 +23,7 @@ urlpatterns = patterns(
     url(r'^create/$', views.create_proposal, name='proposal-create'),
     url(r'^to_review/$', views.proposals_to_review, name='proposals-to-review'),
     url(r'^(?P<slug>[\w-]+)/$', views.detail_proposal, name='proposal-detail'),
+    url(r'^(?P<slug>[\w-]+)~(?P<hashid>.*)/$', views.detail_proposal, name='proposal-detail'),
     url(r'^(?P<slug>[\w-]+)/delete/$', views.delete_proposal, name='proposal-delete'),
     url(r'^(?P<slug>[\w-]+)/update/$', views.update_proposal, name='proposal-update'),
     url(r'^(?P<slug>[\w-]+)/upload-content/$', views.proposal_upload_content, name='proposal-upload-content'),